# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stories API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/stories/{story_id}/story-items": {
            "get": {
                "tags": [
                    "Story"
                ],
                "summary": "Получение списка элементов сториса",
                "operationId": "ba103a4ef750aecff94f50fa0c116bb0",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Номер телефона",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "7000801"
                    },
                    {
                        "name": "story_id",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Связанные данные\n * Доступные связи: 'storyItemButtons'. Полученные вложенные данные будут отсортированы по полю - position в порядке возрастания, также is_active = true.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "storyItemButtons"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Сортировка данных\n * Сортируемые поля: 'title', 'position'. Если передать значение без знака, отсортирует в порядке возрастания, если со знаком минус (-title), отсортирует в порядке убывания.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "position"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID элемента сториса",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": ""
                    },
                    {
                        "name": "is_published",
                        "in": "query",
                        "description": "Опубликован\n * Информация о том, опубликован ли сторис, или нет.",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "element_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Тестовый Сторис"
                                                    },
                                                    "media_type": {
                                                        "type": "string",
                                                        "example": "media_file"
                                                    },
                                                    "file_path": {
                                                        "type": "string",
                                                        "example": "http://stories.test/storage/stories/items/01J7BMT94V1W2D0MFYPFWEAWEB.webp"
                                                    },
                                                    "link": {
                                                        "type": "string",
                                                        "example": "https://www.instagram.com/p/CRYgfzNnoST/?utm_medium=copy_link"
                                                    },
                                                    "position": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "is_published": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/stories/{story_id}/story-items/{story_item_id}": {
            "get": {
                "tags": [
                    "Story"
                ],
                "summary": "Получение определенного элемента сториса через story_item_id",
                "operationId": "1a5743691b2c615fca376bbd699ed32e",
                "parameters": [
                    {
                        "name": "story_id",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    },
                    {
                        "name": "story_item_id",
                        "in": "path",
                        "description": "ID элемента сториса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 21
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Номер телефона",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "7000801"
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Связанные данные\n * Доступные связи: 'storyItemButtons'. Полученные вложенные данные будут отсортированы по полю - position в порядке возрастания, также is_active = true.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "storyItemButtons"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "element_id": {
                                                    "type": "integer",
                                                    "example": 21
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Тестовый Элемент Сториса"
                                                },
                                                "media_type": {
                                                    "type": "string",
                                                    "example": "media_file"
                                                },
                                                "file_path": {
                                                    "type": "string",
                                                    "example": "http://stories.test/storage/stories/items/01J7BMT94V1W2D0MFYPFWEAWEB.webp"
                                                },
                                                "link": {
                                                    "type": "string",
                                                    "example": "https://www.instagram.com/p/CRYgfzNnoST/?utm_medium=copy_link"
                                                },
                                                "position": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "is_published": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "name": "apiKey",
                "in": "header"
            }
        }
    },
...
}